feedback

+++ codex-rs/tui/src/common.rs
@@ -1,1 +1,1 @@
-pub(crate) const DEFAULT_WRAP_COLS: usize = 80;
+pub(crate) const DEFAULT_WRAP_COLS: u16 = 80;
